Refactor galaxy.jobs.datasets into galaxy.job_execution.datasets.
Needed for metadata evaluation stuff (at least in testing) and removes some circular dependencies in the form of stuff in galaxy.tools depending on stuff in galaxy.jobs.
