+++ lib/galaxy/tools/actions/metadata.py
@@ -2,7 +2,7 @@ import logging
 import os
 from json import dumps
 
-from galaxy.jobs.datasets import DatasetPath
+from galaxy.job_execution.datasets import DatasetPath
 from galaxy.metadata import get_metadata_compute_strategy
 from galaxy.util.odict import odict
 from . import ToolAction
