+++ test/unit/jobs/test_datasets.py
@@ -1,4 +1,4 @@
-from galaxy.jobs.datasets import DatasetPath
+from galaxy.job_execution.datasets import DatasetPath
 
 
 def test_dataset_path():
